setup jwtservice authentication in controllers

+++ src/main/java/dev/daycareworkers/controllers/GradeController.java
@@ -1,7 +1,12 @@
 package dev.daycareworkers.controllers;
 
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.interfaces.DecodedJWT;
 import dev.daycareworkers.entities.Grade;
+import dev.daycareworkers.exceptions.InvalidJWTException;
+import dev.daycareworkers.exceptions.UnauthorizedUserException;
 import dev.daycareworkers.services.GradeService;
+import dev.daycareworkers.services.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,24 +20,53 @@ public class GradeController {
 
     @Autowired
     GradeService gradeService;
+    @Autowired
+    JwtService jwtService;
 
     @PostMapping("/grades")
     @ResponseBody
-    public ResponseEntity<Grade> registerGrade(@RequestBody Grade grade) {
-        return new ResponseEntity<Grade>(this.gradeService.registerGrade(grade), HttpStatus.CREATED);
+    public ResponseEntity<Grade> registerGrade(@RequestBody Grade grade, @RequestHeader("auth") String jwt) {
+        if (jwtService.validateJWT(jwt)) {
+            DecodedJWT decodedJWT = JWT.decode(jwt);
+            String role = decodedJWT.getClaim("role").asString();
+
+            if (role.equals("teacher")) {
+                return new ResponseEntity<Grade>(this.gradeService.registerGrade(grade), HttpStatus.CREATED);
+            }
+            throw new UnauthorizedUserException();
+        }
+        throw new InvalidJWTException();
+
     }
 
     @GetMapping("/grades/{id}")
     @ResponseBody
-    public List<Grade> gradesBySid(@PathVariable int id) {
-        //int sid = Integer.parseInt(id);
-        return this.gradeService.findGradesBySid(id);
+    public List<Grade> gradesBySid(@PathVariable int id, @RequestHeader("auth") String jwt) {
+        if (jwtService.validateJWT(jwt)) {
+            DecodedJWT decodedJWT = JWT.decode(jwt);
+            String role = decodedJWT.getClaim("role").asString();
+
+            //int sid = Integer.parseInt(id);
+            return this.gradeService.findGradesBySid(id);
+
+            throw new UnauthorizedUserException();
+        }
+        throw new InvalidJWTException();
     }
 
     @DeleteMapping("/grades/{id}")
     @ResponseBody
-    public boolean deleteGradeById(@PathVariable String id){
-        int gid = Integer.parseInt(id);
-        return this.gradeService.deleteGradeById(gid);
+    public boolean deleteGradeById(@PathVariable String id, @RequestHeader("auth") String jwt){
+        if (jwtService.validateJWT(jwt)) {
+            DecodedJWT decodedJWT = JWT.decode(jwt);
+            String role = decodedJWT.getClaim("role").asString();
+
+            if (role.equals("teacher")) {
+                int gid = Integer.parseInt(id);
+                return this.gradeService.deleteGradeById(gid);
+            }
+            throw new UnauthorizedUserException();
+        }
+        throw new InvalidJWTException();
     }
 }
